More extreme revamping of connection logic
Bluetooth/Wifi Client/Server are all combined now.  A lot of the logic
for the connection service is shared now too.
