+++ CardGames/src/com/worthwhilegames/cardgames/shared/bluetooth/BluetoothConnectionService.java
@@ -1,25 +1,17 @@
 package com.worthwhilegames.cardgames.shared.bluetooth;
 
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-
-import org.json.JSONException;
-import org.json.JSONObject;
 
 import android.bluetooth.BluetoothAdapter;
 import android.bluetooth.BluetoothDevice;
 import android.bluetooth.BluetoothSocket;
 import android.content.Context;
-import android.os.Bundle;
 import android.os.Handler;
-import android.os.Message;
 import android.util.Log;
 
 import com.worthwhilegames.cardgames.shared.Util;
-import com.worthwhilegames.cardgames.shared.connection.ConnectionConstants;
-import com.worthwhilegames.cardgames.shared.connection.IConnectionService;
-import com.worthwhilegames.cardgames.shared.connection.ISocket;
+import com.worthwhilegames.cardgames.shared.connection.ConnectionService;
+import com.worthwhilegames.cardgames.shared.connection.IConnectThread;
 
 /**
  * This class does all the work for setting up and managing Bluetooth
@@ -27,7 +19,7 @@ import com.worthwhilegames.cardgames.shared.connection.ISocket;
  * incoming connections, a thread for connecting with a device, and a
  * thread for performing data transmissions when connected.
  */
-public class BluetoothConnectionService implements IConnectionService {
+public class BluetoothConnectionService extends ConnectionService {
 	/**
 	 * The Logcat Debug tag
 	 */
@@ -39,99 +31,14 @@ public class BluetoothConnectionService implements IConnectionService {
 	private final BluetoothAdapter mAdapter;
 
 	/**
-	 * The Handler to post messages to when something changes in this service
-	 */
-	private final Handler mHandler;
-
-	/**
-	 * The Thread that handles the process of starting a connection
-	 */
-	private ConnectThread mConnectThread;
-
-	/**
-	 * The Thread that handles the sending/receiving of data once
-	 * a connection has been established
-	 */
-	private ConnectedThread mConnectedThread;
-
-	/**
-	 * The current state of this Bluetooth connection
-	 * 
-	 * See ConnectionConstants.STATE_* for possible values
-	 */
-	private int mState;
-
-	/**
-	 * The MAC address of the remote device this service is connected to
-	 */
-	private String deviceAddress;
-
-	/**
 	 * Constructor. Initializes information needed to create a Bluetooth connection
 	 * 
 	 * @param context  The UI Activity Context
 	 * @param handler  A Handler to send messages back to the UI Activity
 	 */
 	public BluetoothConnectionService(Context context, Handler handler) {
+		super(context, handler);
 		mAdapter = BluetoothAdapter.getDefaultAdapter();
-		mState = ConnectionConstants.STATE_NONE;
-		mHandler = handler;
-	}
-
-	/**
-	 * Set the current state of the Bluetooth connection
-	 * 
-	 * @param state  An integer defining the current connection state
-	 */
-	private synchronized void setState(int state) {
-		if (Util.isDebugBuild()) {
-			Log.d(TAG, "setState() " + mState + " -> " + state);
-		}
-
-		mState = state;
-
-		// Send a message to the Handler letting them know the state has been updated
-		Message msg = mHandler.obtainMessage(BluetoothConstants.STATE_MESSAGE, -1, -1);
-		Bundle bundle = new Bundle();
-		bundle.putInt(ConnectionConstants.KEY_STATE_MESSAGE, state);
-		bundle.putString(ConnectionConstants.KEY_DEVICE_ID, deviceAddress);
-		msg.setData(bundle);
-		msg.sendToTarget();
-	}
-
-	/**
-	 * Return the current connection state.
-	 * 
-	 * @return the current state
-	 */
-	@Override
-	public synchronized int getState() {
-		return mState;
-	}
-
-	/**
-	 * Start the Bluetooth service. Cancels any currently connected connections,
-	 * and sets the state to STATE_LISTEN.
-	 */
-	@Override
-	public synchronized void start() {
-		if (Util.isDebugBuild()) {
-			Log.d(TAG, "start");
-		}
-
-		// Cancel any thread attempting to make a connection
-		if (mConnectThread != null) {
-			mConnectThread.cancel();
-			mConnectThread = null;
-		}
-
-		// Cancel any thread currently running a connection
-		if (mConnectedThread != null) {
-			mConnectedThread.cancel();
-			mConnectedThread = null;
-		}
-
-		setState(ConnectionConstants.STATE_LISTEN);
 	}
 
 	/**
@@ -141,23 +48,7 @@ public class BluetoothConnectionService implements IConnectionService {
 	 */
 	@Override
 	public synchronized void connect(String name) {
-		if (Util.isDebugBuild()) {
-			Log.d(TAG, "connect to: " + name);
-		}
-
-		// Cancel any thread attempting to make a connection
-		if (mState == ConnectionConstants.STATE_CONNECTING) {
-			if (mConnectThread != null) {
-				mConnectThread.cancel();
-				mConnectThread = null;
-			}
-		}
-
-		// Cancel any thread currently running a connection
-		if (mConnectedThread != null) {
-			mConnectedThread.cancel();
-			mConnectedThread = null;
-		}
+		super.connect(name);
 
 		BluetoothAdapter bluetoothAdapter = BluetoothAdapter.getDefaultAdapter();
 		BluetoothDevice device = bluetoothAdapter.getRemoteDevice(name);
@@ -165,104 +56,15 @@ public class BluetoothConnectionService implements IConnectionService {
 		// Start the thread to connect with the given device
 		mConnectThread = new ConnectThread(device);
 		mConnectThread.start();
-
-		setState(ConnectionConstants.STATE_CONNECTING);
 	}
 
-	/**
-	 * Start the ConnectedThread to begin managing a Bluetooth connection.
-	 * 
-	 * Called once a Bluetooth connection has been established (either by ConnectThread
-	 * or AcceptThread).
-	 * 
-	 * @param socket  The ISocket on which the connection was made
-	 */
-	@Override
-	public synchronized void connected(ISocket socket) {
-		if (Util.isDebugBuild()) {
-			Log.d(TAG, "connected Socket");
-		}
-		com.worthwhilegames.cardgames.shared.bluetooth.BluetoothSocket actualSocket = (com.worthwhilegames.cardgames.shared.bluetooth.BluetoothSocket) socket;
-
-		// Cancel the thread that completed the connection
-		if (mConnectThread != null) {
-			mConnectThread.cancel();
-			mConnectThread = null;
-		}
-
-		// Cancel any thread currently running a connection
-		if (mConnectedThread != null) {
-			mConnectedThread.cancel();
-			mConnectedThread = null;
-		}
-
-		// Start the thread to manage the connection and perform transmissions
-		mConnectedThread = new ConnectedThread(actualSocket.getSocket());
-		mConnectedThread.start();
-
-		// Store the remote device's address
-		deviceAddress = socket.toString();
-
-		// Update our state
-		setState(ConnectionConstants.STATE_CONNECTED);
-	}
-
-	/**
-	 * Stop all connections, and set the state to STATE_NONE
-	 */
-	@Override
-	public synchronized void stop() {
-		if (Util.isDebugBuild()) {
-			Log.d(TAG, "stop");
-		}
-
-		if (mConnectThread != null) {
-			mConnectThread.cancel();
-			mConnectThread = null;
-		}
-
-		if (mConnectedThread != null) {
-			mConnectedThread.cancel();
-			mConnectedThread = null;
-		}
-
-		setState(ConnectionConstants.STATE_NONE);
-	}
-
-	/**
-	 * Write to the ConnectedThread in an unsynchronized manner
-	 * 
-	 * @param out The bytes to write
-	 * @see ConnectedThread#write(byte[])
-	 */
-	@Override
-	public void write(byte[] out) {
-		// Create temporary object
-		ConnectedThread r;
-
-		// Synchronize a copy of the ConnectedThread
-		synchronized (this) {
-			if (mState != ConnectionConstants.STATE_CONNECTED) {
-				return;
-			}
-
-			r = mConnectedThread;
-		}
-
-		if (Util.isDebugBuild()) {
-			Log.d(TAG, "msgsent: " + new String(out));
-		}
-
-		// Perform the write unsynchronized
-		r.write(out);
-	}
 
 	/**
 	 * This thread runs while attempting to make an outgoing connection
 	 * with a device. It runs straight through; the connection either
 	 * succeeds or fails.
 	 */
-	private class ConnectThread extends Thread {
+	private class ConnectThread extends Thread implements IConnectThread {
 		/**
 		 * The BluetoothSocket this connection will be opened on
 		 */
@@ -349,123 +151,6 @@ public class BluetoothConnectionService implements IConnectionService {
 
 			// Start the connected thread
 			connected(new com.worthwhilegames.cardgames.shared.bluetooth.BluetoothSocket(mmSocket));
-		}
-
-		/**
-		 * Cancel the current operation
-		 */
-		public void cancel() {
-			try {
-				mmSocket.close();
-			} catch (IOException e) {
-				Log.e(TAG, "close() of connect socket failed", e);
-			}
-		}
-	}
-
-	/**
-	 * This thread runs during a connection with a remote device.
-	 * It handles all incoming and outgoing transmissions.
-	 */
-	private class ConnectedThread extends Thread {
-		/**
-		 * The BluetoothSocket that messages are sent/received on
-		 */
-		private final BluetoothSocket mmSocket;
-
-		/**
-		 * The InputStream to read messages from
-		 */
-		private final InputStream mmInStream;
-
-		/**
-		 * The OutputStream to write messages to
-		 */
-		private final OutputStream mmOutStream;
-
-		/**
-		 * Create a new ConnectedThread with the given socket
-		 * 
-		 * @param socket the socket to use for this thread
-		 */
-		public ConnectedThread(BluetoothSocket socket) {
-			if (Util.isDebugBuild()) {
-				Log.d(TAG, "create ConnectedThread");
-			}
-
-			mmSocket = socket;
-			InputStream tmpIn = null;
-			OutputStream tmpOut = null;
-
-			// Get the BluetoothSocket input and output streams
-			try {
-				tmpIn = socket.getInputStream();
-				tmpOut = socket.getOutputStream();
-			} catch (IOException e) {
-				Log.e(TAG, "temp sockets not created", e);
-			}
-
-			mmInStream = tmpIn;
-			mmOutStream = tmpOut;
-		}
-
-		/* (non-Javadoc)
-		 * @see java.lang.Thread#run()
-		 */
-		@Override
-		public void run() {
-			if (Util.isDebugBuild()) {
-				Log.i(TAG, "BEGIN mConnectedThread");
-			}
-
-			byte[] buffer = new byte[1024];
-			int bytes;
-
-			// Keep listening to the InputStream while connected
-			while (true) {
-				try {
-					// Read from the InputStream
-					bytes = mmInStream.read(buffer);
-
-					JSONObject obj = new JSONObject(new String(buffer, 0, bytes));
-
-					if (Util.isDebugBuild()) {
-						Log.d(TAG, "msgrx: " + obj.toString());
-					}
-
-					// Send the obtained bytes to the UI Activity
-					Message msg = mHandler.obtainMessage(BluetoothConstants.READ_MESSAGE, -1, -1, null);
-					Bundle data = new Bundle();
-					data.putInt(ConnectionConstants.KEY_MESSAGE_TYPE, obj.getInt(ConnectionConstants.KEY_MESSAGE_TYPE));
-					if (obj.has(ConnectionConstants.KEY_MSG_DATA)) {
-						data.putString(ConnectionConstants.KEY_MESSAGE_RX, obj.get(ConnectionConstants.KEY_MSG_DATA).toString());
-					}
-					data.putString(ConnectionConstants.KEY_DEVICE_ID, deviceAddress);
-					msg.setData(data);
-					msg.sendToTarget();
-				} catch (IOException e) {
-					Log.e(TAG, "disconnected", e);
-
-					// Start the service over to restart listening mode
-					BluetoothConnectionService.this.start();
-					break;
-				} catch (JSONException e) {
-					e.printStackTrace();
-				}
-			}
-		}
-
-		/**
-		 * Write to the connected OutStream
-		 * 
-		 * @param buffer The bytes to write
-		 */
-		public void write(byte[] buffer) {
-			try {
-				mmOutStream.write(buffer);
-			} catch (IOException e) {
-				Log.e(TAG, "Exception during write", e);
-			}
 		}
 
 		/**
