+++ CardGames/src/com/worthwhilegames/cardgames/shared/bluetooth/BluetoothConstants.java
@@ -17,24 +17,4 @@ public class BluetoothConstants {
 	 * will be connecting from.
 	 */
 	public static final UUID MY_UUID = UUID.fromString("9d6b7fe4-d2cd-37f9-950b-0aad096c2d57");
-
-	/**
-	 *  Name for the SDP record when creating server socket
-	 */
-	public static final String SOCKET_NAME = "CardGames";
-
-	/**
-	 * The constant used internally from the BluetoothConnectionService
-	 * to the BluetoothClient/BluetoothServer indicating that we read a message
-	 * from the Bluetooth port.
-	 */
-	protected static final int READ_MESSAGE  = 0;
-
-	/**
-	 * THe constant used internally from the BluetoothConnectionService
-	 * to the BluetoothClient/BluetoothServer indicating that the state of the
-	 * Bluetooth connection has changed.
-	 */
-	protected static final int STATE_MESSAGE = 1;
-
 }
