+++ CardGames/src/com/worthwhilegames/cardgames/shared/connection/ISocket.java
@@ -1,6 +1,8 @@
 package com.worthwhilegames.cardgames.shared.connection;
 
 import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 
 /**
  * Abstract out a regular Socket.
@@ -17,4 +19,33 @@ public interface ISocket {
 	 */
 	void close() throws IOException;
 
+	/**
+	 * Get the input stream for this socket
+	 * 
+	 * @return the input stream for this socket
+	 * @throws IOException
+	 */
+	InputStream getInputStream() throws IOException;
+
+	/**
+	 * Get the output stream for this socket
+	 * 
+	 * @return the output stream for this socket
+	 * @throws IOException
+	 */
+	OutputStream getOutputStream() throws IOException;
+
+	/**
+	 * Stop all output for this socket
+	 * 
+	 * @throws IOException
+	 */
+	void shutdownOutput() throws IOException;
+
+	/**
+	 * Stop all input for this socket
+	 * 
+	 * @throws IOException
+	 */
+	void shutdownInput() throws IOException;
 }
