+++ CardGames/src/com/worthwhilegames/cardgames/shared/wifi/WifiSocket.java
@@ -1,6 +1,8 @@
 package com.worthwhilegames.cardgames.shared.wifi;
 
 import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 import java.net.Socket;
 
 import com.worthwhilegames.cardgames.shared.connection.ISocket;
@@ -27,6 +29,38 @@ public class WifiSocket implements ISocket {
 	@Override
 	public void close() throws IOException {
 		mSocket.close();
+	}
+
+	/* (non-Javadoc)
+	 * @see com.worthwhilegames.cardgames.shared.connection.ISocket#getInputStream()
+	 */
+	@Override
+	public InputStream getInputStream() throws IOException {
+		return mSocket.getInputStream();
+	}
+
+	/* (non-Javadoc)
+	 * @see com.worthwhilegames.cardgames.shared.connection.ISocket#getOutputStream()
+	 */
+	@Override
+	public OutputStream getOutputStream() throws IOException {
+		return mSocket.getOutputStream();
+	}
+
+	/* (non-Javadoc)
+	 * @see com.worthwhilegames.cardgames.shared.connection.ISocket#shutdownOutput()
+	 */
+	@Override
+	public void shutdownOutput() throws IOException {
+		mSocket.shutdownOutput();
+	}
+
+	/* (non-Javadoc)
+	 * @see com.worthwhilegames.cardgames.shared.connection.ISocket#shutdownInput()
+	 */
+	@Override
+	public void shutdownInput() throws IOException {
+		mSocket.shutdownInput();
 	}
 
 	/* (non-Javadoc)
